+++ fast_carpenter/backends/coffea.py
@@ -4,8 +4,8 @@ Functions to run a job using Coffea
 import copy
 from fast_carpenter.masked_tree import MaskedUprootTree
 from collections import namedtuple
-from coffea import processor
-from coffea.processor import futures_executor, run_uproot_job
+from coffea import processor as cop
+import logging
 
 
 EventRanger = namedtuple("EventRanger", "start_entry stop_entry entries_in_block")
@@ -14,7 +14,7 @@ ChunkConfig = namedtuple("ChunkConfig", "dataset")
 ConfigProxy = namedtuple("ConfigProxy", "name eventtype")
 
 
-class stages_accumulator(processor.AccumulatorABC):
+class stages_accumulator(cop.AccumulatorABC):
     def __init__(self, stages):
         self._zero = copy.deepcopy(stages)
         self._value = copy.deepcopy(stages)
@@ -32,13 +32,13 @@ class stages_accumulator(processor.AccumulatorABC):
             stage.merge(other[i])
 
 
-class FASTProcessor(processor.ProcessorABC):
+class FASTProcessor(cop.ProcessorABC):
     def __init__(self, sequence):
 
         self._columns = list()
         self._sequence = sequence
-        accumulator_dict = {'stages': processor.dict_accumulator({})}
-        self._accumulator = processor.dict_accumulator(accumulator_dict)
+        accumulator_dict = {'stages': cop.dict_accumulator({})}
+        self._accumulator = cop.dict_accumulator(accumulator_dict)
 
     @property
     def columns(self):
@@ -82,14 +82,86 @@ class FASTProcessor(processor.ProcessorABC):
         return accumulator
 
 
+def load_execution_cfg(config):
+    if not isinstance(config, str):
+        return config
+
+    import yaml
+    with open(config, "r") as infile:
+        cfg = yaml.safe_load(infile)
+        return cfg
+
+
+def configure_parsl(n_threads, monitoring, **kwargs):
+    from parsl.config import Config
+    from parsl.executors.threads import ThreadPoolExecutor
+    from parsl.addresses import address_by_hostname
+
+    if monitoring:
+        from parsl.monitoring import MonitoringHub
+        monitoring = MonitoringHub(hub_address=address_by_hostname(),
+                                   hub_port=55055,
+                                   logging_level=logging.INFO,
+                                   resource_monitoring_interval=10,
+                                   )
+    else:
+        monitoring = None
+
+    local_threads = ThreadPoolExecutor(max_threads=n_threads,
+                                       label='local_threads')
+    config = Config(executors=[local_threads],
+                    monitoring=monitoring,
+                    strategy=None,
+                    app_cache=True,
+                    )
+    return config
+
+
+def configure_dask(client, **kwargs):
+    from dask.distributed import Client
+
+    if client:
+        return Client(client)
+    else:
+        return Client(**kwargs)
+
+
+def create_executor(args):
+    exe_type = args.mode.split(":", 1)[-1].lower()
+    exe_args = {}
+    if getattr(args, "execution_cfg", None):
+        exe_args = load_execution_cfg(args.execution_cfg)
+
+    if exe_type == "local":
+        executor = cop.futures_executor
+        exe_args.setdefault('workers', args.ncores)
+        exe_args.setdefault('flatten', False)
+    elif exe_type == "parsl":
+        executor = cop.parsl_executor
+        exe_args.setdefault('n_threads', args.ncores)
+        exe_args.setdefault('monitoring', False)
+        if 'config' not in exe_args:
+            exe_args['config'] = configure_parsl(**exe_args)
+        exe_args.setdefault('flatten', False)
+    elif exe_type == "dask":
+        executor = cop.dask_executor
+        exe_args.setdefault('processes', False)
+        exe_args.setdefault('threads_per_worker', 2)
+        exe_args.setdefault('n_workers', 2)
+        exe_args.setdefault('memory_limit', '1GB')
+        exe_args.setdefault('client', None)
+        exe_args['client'] = configure_dask(**exe_args)
+    else:
+        msg = "Coffea executor not yet included in fast-carpenter: '%s'"
+        raise NotImplementedError(msg % exe_type)
+
+    return executor, exe_args
+
+
 def execute(sequence, datasets, args):
     fp = FASTProcessor(sequence)
 
-    executor = futures_executor
-    exe_args = {'workers': args.ncores,
-                'chunksize': args.blocksize,
-                'maxchunks': args.nblocks_per_dataset,
-                'function_args': {'flatten': False}}
+    executor, exe_args = create_executor(args)
 
     coffea_datasets = {}
     for ds in datasets:
@@ -97,5 +169,11 @@ def execute(sequence, datasets, args):
         coffea_datasets[ds.name].pop('name')
         coffea_datasets[ds.name]['treename'] = coffea_datasets[ds.name].pop('tree')
 
-    out = run_uproot_job(coffea_datasets, 'events', fp, executor, executor_args=exe_args)
+    maxchunks = args.nblocks_per_dataset
+    if maxchunks < 1:
+        maxchunks = None
+    out = cop.run_uproot_job(coffea_datasets, 'events', fp,
+                             executor, executor_args=exe_args,
+                             chunksize=args.blocksize, maxchunks=maxchunks)
+
     return out["stages"], out["results"]